feat: moved buttons to ButtonSection
